pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "graphrag-lab"
version = "0.1.0"
description = "M1 GraphRAG research scaffold with a CPU-safe toy pipeline."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = []

[project.scripts]
graphrag-m1 = "graphrag_lab.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
